Refactor generative_model_tool to tools/core/llms
